ecpay: fold entrypoint into api.go and delete server.go; drop unused .devcontainer
- ecpay: move main(), logging setup, Prometheus metrics and the mysql driver
  import from server.go into api.go, then delete server.go. After removing the
  dead gRPC path there was no reason to keep a separate file — api.go is now the
  whole service. Payment e2e (scenario 2) passes on a rebuilt image.
- README: fix the ecpay row (Go (gRPC) → Go (Gin)); ecpay README layout drops
  server.go.
- Remove .devcontainer/devcontainer.json — nothing in the build, CI, or Taskfile
  references it (only VS Code Dev Containers / Codespaces would), so it was dead
  config.

Note: ecfront/package-lock.json is intentionally kept — CI's `npm ci` requires it
for reproducible installs.
