--- .devcontainer/devcontainer.json
@@ -1,6 +1,0 @@
-{
-    "image": "mcr.microsoft.com/devcontainers/universal:2",
-    "features": {
-    },
-    "postCreateCommand": "go install github.com/go-task/task/v3/cmd/task@latest && cd ecfront && npm install"
-  }
